pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "simargl"
version = "0.1.0"
description = "Task-to-code retrieval — MCP server and web UI"
requires-python = ">=3.10"
readme = {file = "README.md", content-type = "text/markdown"}
dependencies = [
    "sentence-transformers>=2.2",
    "mcp>=1.0",
    "numpy>=1.24",
    "tqdm>=4.60",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/szholobetsky/simargl"
Repository = "https://github.com/szholobetsky/simargl"

[project.optional-dependencies]
ui      = ["gradio>=4.0"]
postgres = ["psycopg2-binary>=2.9", "pgvector>=0.2"]
http    = ["uvicorn>=0.24"]
ingest  = ["gitpython>=3.1", "requests>=2.28", "beautifulsoup4>=4.12"]
selenium = ["selenium>=4.0"]

[project.scripts]
simargl     = "simargl.ui.cli:main"
simargl-mcp = "simargl.mcp_server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["simargl*"]
